add detection for 'MergeConflicts'
This EA smell occurs when multiple source elements are are having concerns for the sameelement,
i.e. there are multiple incoming relationships to the same element.
This can lead to conflicting changes made in parallel by the source elements,
and an unnecessary resolution of the conflict needs to be done. If no conflicting change is present,
having multiple source elements responsible the same element could still indicate in redundancy in the source elements.