add detection for 'RedundantRelationship'
The EA smell occurs if there is an element that has multiple incoming relationships,
that is targeted by the same source element, through indirect paths in the model.
This is a special case of the EA smell Merge Conflict, as both these EA smells will have multiple incoming relationships,
but RedundantRelationship is targeted by the same element.